Move all the AR indexers to catalog/indexers.
