+++ bika/health/catalog/indexers/analysisrequest.py
@@ -9,10 +9,56 @@ from bika.lims import api
 from bika.lims import logger
 from bika.lims.interfaces import IAnalysisRequest
 from plone.indexer import indexer
+from bika.lims.interfaces import IBikaCatalog
 from bika.lims.interfaces import IBikaCatalogAnalysisRequestListing
 
 
+# Defining the indexes for this extension. Since this is an extension, no
+# getter is created so we need to create indexes in that way.
+# TODO-catalog: delete this index
+@indexer(IAnalysisRequest, IBikaCatalog)
+def getPatientUID(instance):
+    field = instance.getField('Patient', '')
+    item = field.get(instance) if field else None
+    value = item and item.UID() or ''
+    return value
+
+
 @indexer(IAnalysisRequest, IBikaCatalogAnalysisRequestListing)
+def getPatientUID(instance):
+    field = instance.getField('Patient', '')
+    item = field.get(instance) if field else None
+    value = item and item.UID() or ''
+    return value
+
+
+@indexer(IAnalysisRequest, IBikaCatalogAnalysisRequestListing)
+def getDoctorUID(instance):
+    field = instance.getField('Doctor', '')
+    item = field.get(instance) if field else None
+    value = item and item.UID() or ''
+    return value
+
+
+# We use this index to sort columns and filter lists
+@indexer(IAnalysisRequest, IBikaCatalogAnalysisRequestListing)
+def getPatient(instance):
+    field = instance.getField('Patient', '')
+    item = field.get(instance) if field else None
+    value = item and item.Title() or ''
+    return value
+
+
+# We use this index to sort columns and filter lists
+@indexer(IAnalysisRequest, IBikaCatalogAnalysisRequestListing)
+def getPatientID(instance):
+    field = instance.getField('Patient', '')
+    item = field.get(instance) if field else None
+    value = item and item.getId() or ''
+    return value
+
+
+@indexer(IAnalysisRequest)
 def listing_searchable_text(instance):
     """ Indexes values of desired fields for searches in listing view. All the
     field names added to 'plain_text_fields' will be available to search by
@@ -25,7 +71,10 @@ def listing_searchable_text(instance):
     """
     entries = []
     plain_text_fields = ('getId', 'getContactFullName', 'getSampleTypeTitle',
-                         'getSamplePointTitle')
+                         'getSamplePointTitle', 'getCreatorFullName',
+                         'getProfilesTitle', 'getStorageLocationTitle',
+                         'getClientOrderNumber', 'getClientReference',
+                         'getClientSampleID', 'getTemplateTitle', )
 
     # Concatenate plain text fields as they are
     for field_name in plain_text_fields:
@@ -35,9 +84,11 @@ def listing_searchable_text(instance):
             logger.error("{} has no attribute called '{}' ".format(
                             repr(instance), field_name))
             continue
+        if isinstance(value, list):
+            value = " ".join(value)
         entries.append(value)
 
-    # Getter of senaite.health extension fields are not created. That's why
+    # Getters of senaite.health extension fields are not created. That's why
     # we are adding them manually
     patient_field = instance.getField('Patient', '')
     patient = patient_field.get(instance) if patient_field else None
