+++ bika/health/catalog/indexers/configure.zcml
@@ -1,8 +1,13 @@
 <configure
     xmlns="http://namespaces.zope.org/zope"
     i18n_domain="bika">
-  <unconfigure>
-      <adapter name="listing_searchable_text" factory=".analysisrequest.listing_searchable_text" />
-  </unconfigure>
-  <adapter name="listing_searchable_text" factory=".analysisrequest.listing_searchable_text"/>
+
+    <unconfigure>
+        <adapter name="listing_searchable_text" factory=".analysisrequest.listing_searchable_text" />
+    </unconfigure>
+    <adapter name="listing_searchable_text" factory=".analysisrequest.listing_searchable_text"/>
+    <adapter factory=".analysisrequest.getPatientUID" name="getPatientUID" />
+    <adapter factory=".analysisrequest.getPatient" name="getPatient" />
+    <adapter factory=".analysisrequest.getPatientID" name="getPatientID" />
+    <adapter factory=".analysisrequest.getDoctorUID" name="getDoctorUID" />
 </configure>
